Add TranslatableComponent and rearrange translation methods
